expand range of cocktail dependency to include current version
With npm 3, this will allow a single copy of cocktail.js to be installed when a project depends on both cocktail and this module, due to the "maximally flat" algorithm used by npm 3. 

Currently using this trait requires an additional older copy of cocktail to be loaded as an additional dependency for projects.